% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_retracted.R
\name{count_retracted}
\alias{count_retracted}
\title{count_retracted}
\usage{
count_retracted(df)
}
\arguments{
\item{df}{A dataframe with an "is_retracted" column indicating whether a reference has been retracted}
}
\value{
Text reporting the number of retracted references and total references
}
\description{
count_retracted
}
\examples{
Add later
}
