% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_retracted.R
\name{check_retracted}
\alias{check_retracted}
\title{check_retracted}
\usage{
check_retracted(refs)
}
\arguments{
\item{refs}{A dataframe object containing bibliographic references.}
}
\value{
A dataframe object of references with a column indicating whether the article has been retracted (according to the RetractionWatch database).
}
\description{
check_retracted
}
\examples{
Add later.
}
