% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runShiny.R
\name{runShiny}
\alias{runShiny}
\title{A wrapper function to run the Shiny App}
\usage{
runShiny(app = "RetractionFindR")
}
\arguments{
\item{app}{RetractionFindR}
}
\value{
A shiny app
}
\description{
A wrapper function to run the Shiny App
}
