% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separate_retracted.R
\name{separate_retracted}
\alias{separate_retracted}
\title{separate_retracted}
\usage{
separate_retracted(df)
}
\arguments{
\item{df}{A dataframe object of bibliographic data, including column with a binary indicator of whether the article has been retracted (according to the RetractionWatch database).}
}
\value{
A list of dataframes, one containing only retracted articles and one containing only non-retracted ones.
}
\description{
Separates a list of references into retracted and non-retracted articles.
}
\examples{
Add later
}
